Add MD extension to SmallSort challenge (#60)
The SmallSort challenge file is missing the `.md` extension, so our book
renders an empty page (see
[here](https://model-checking.github.io/verify-rust-std/challenges/0008-smallsort.html)).
This PR adds the `.md` extension. I tested locally with `mdbook serve`
to ensure it renders correctly.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.