Improve performance for align offset harness (#89)
Currently, this harness takes 21 minutes in CI. The only point of this
harness is to verify the contract for pointee types with a non-power of
two byte size--17 was an arbitrary choice. Reducing to 5 and changing
the solver reduces verification time to 57 seconds on my local machine.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.